# tools/WillowWSUP/WsupApi/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Willow.Support.Api",
    "version": "v1"
  },
  "paths": {
...
    "/CustomerInstances/GetAll": {
      "get": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_GetAll",
        "parameters": [
          {
            "name": "includeApplications",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerInstance"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/CustomerInstances/Get": {
      "get": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_Get",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "includeApplications",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInstance"
                }
              }
            }
          }
        }
      }
    },
    "/CustomerInstances": {
      "post": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_Create",
        "requestBody": {
          "x-name": "customerInstance",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInstance"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInstance"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_Update",
        "requestBody": {
          "x-name": "customerInstance",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerInstance"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CustomerInstances"
        ],
        "operationId": "CustomerInstances_Delete",
        "parameters": [
          {
            "name": "customerInstanceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/Customers/Get": {
      "get": {
        "tags": [
          "Customers"
        ],
        "operationId": "Customers_Get",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/Customers": {
      "post": {
        "tags": [
          "Customers"
        ],
        "operationId": "Customers_Create",
        "requestBody": {
          "x-name": "customer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Customers"
        ],
        "operationId": "Customers_Update",
        "requestBody": {
          "x-name": "customer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Customers"
        ],
        "operationId": "Customers_Delete",
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/Environments/Get": {
      "get": {
        "tags": [
          "Environments"
        ],
        "operationId": "Environments_Get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Environment"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/Environments": {
      "post": {
        "tags": [
          "Environments"
        ],
        "operationId": "Environments_Create",
        "requestBody": {
          "x-name": "environment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Environment"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Environments"
        ],
        "operationId": "Environments_Update",
        "requestBody": {
          "x-name": "environment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Environment"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Environments"
        ],
        "operationId": "Environments_Delete",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/Regions/Get": {
      "get": {
        "tags": [
          "Regions"
        ],
        "operationId": "Regions_Get",
        "parameters": [
          {
            "name": "shortName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true,
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/Region"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/Regions": {
      "post": {
        "tags": [
          "Regions"
        ],
        "operationId": "Regions_Create",
        "requestBody": {
          "x-name": "region",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Regions"
        ],
        "operationId": "Regions_Update",
        "requestBody": {
          "x-name": "region",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Region"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Region"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Regions"
        ],
        "operationId": "Regions_Delete",
        "parameters": [
          {
            "name": "shortName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Application": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "teamId": {
            "type": "integer",
            "format": "int32"
          },
          "applicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "healthEndpointPath": {
            "type": "string",
            "nullable": true
          },
          "hasPublicEndpoint": {
            "type": "boolean"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "team": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Team"
              }
            ]
          },
          "applicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ApplicationStatus"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          }
        }
      },
...
      "CustomerInstanceApplication": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "applicationId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstanceApplicationStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "application": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Application"
              }
            ]
          },
          "customerInstanceApplicationStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceApplicationStatus"
              }
            ]
          }
        }
      },
      "CustomerInstance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "customerId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "domain": {
            "type": "string",
            "nullable": true
          },
          "fullDomain": {
            "type": "string",
            "nullable": true
          },
          "dnsEnvSuffix": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "fullCustomerInstanceName": {
            "type": "string",
            "nullable": true
          },
          "deploymentPhase": {
            "type": "string",
            "nullable": true
          },
          "lifecycleState": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "azureDigitalTwinsInstance": {
            "type": "string",
            "nullable": true
          },
          "azureDataExplorerInstance": {
            "type": "string",
            "nullable": true
          },
          "stampId": {
            "type": "string",
            "format": "guid"
          },
          "customer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Customer"
              }
            ]
          },
          "customerInstanceStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstanceStatus"
              }
            ]
          },
          "stamp": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Stamp"
              }
            ]
          },
          "customerInstanceApplications": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstanceApplication"
            }
          },
          "buildings": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Building"
            }
          },
          "connectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Connector"
            }
          }
        }
      },
      "Customer": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "salesId": {
            "type": "string",
            "nullable": true
          },
          "customerStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "customerStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerStatus"
              }
            ]
          }
        }
      },
      "CustomerStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "customers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Customer"
            }
          }
        }
      },
      "CustomerInstanceStatus": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "resourceGroupName": {
            "type": "string",
            "nullable": true
          },
          "adxDatabaseUrl": {
            "type": "string",
            "nullable": true
          },
          "adtInstanceUrl": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          }
        }
      },
      "Stamp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "subscriptionId": {
            "type": "string",
            "format": "guid"
          },
          "environmentName": {
            "type": "string",
            "nullable": true
          },
          "regionShortName": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "region": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Region"
              }
            ]
          },
          "environment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          }
        }
      },
      "Region": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "shortName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "customerInstances": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CustomerInstance"
            }
          },
          "stamps": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Stamp"
            }
          }
        }
      },
      "Environment": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "stamps": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Stamp"
            }
          }
        }
      },
      "Building": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
...
      "Connector": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "customerInstanceId": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "connectorStatusId": {
            "type": "integer",
            "format": "int32"
          },
          "connectorTypeId": {
            "type": "string",
            "nullable": true
          },
          "connectorStatus": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorStatus"
              }
            ]
          },
          "connectorType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ConnectorType"
              }
            ]
          },
          "customerInstance": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CustomerInstance"
              }
            ]
          },
          "buildingConnectors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/BuildingConnector"
            }
          }
        }
      },
...
    }
  }
}